# doxi/src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0"
  },
  "basePath" : "/doxi/rest",
...
  "paths" : {
...
    "/pid/cache/size" : {
      "get" : {
        "tags" : [ "PID API" ],
        "summary" : "Returns the current size of the Pid Cache.",
        "description" : "",
        "operationId" : "cacheSize",
        "consumes" : [ "text/plain" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Size found.",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "Size not found."
          }
        }
      }
    },
...
    "/pid/queue/size" : {
      "get" : {
        "tags" : [ "PID API" ],
        "summary" : "Returns the current size of the Pid Queue.",
        "description" : "",
        "operationId" : "queueSize",
        "consumes" : [ "text/plain" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Size found.",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "Size not found."
          }
        }
      }
    },
    "/pid/retrieve" : {
      "get" : {
        "tags" : [ "PID API" ],
        "summary" : "Retrieves a Pid with known ID.",
        "description" : "",
        "operationId" : "retrieve",
        "consumes" : [ "text/plain" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "the ID which should be retrieved",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "PID found.",
            "schema" : {
              "type" : "string"
            }
          },
          "400" : {
            "description" : "PID not found."
          }
        }
      }
    },
...
  }
}